Ensure install/postinstall does not run on self.
Check the install target's package name to determine if that
name is the same as the one in the context of the install event.
If so, then no files need to be copied (and, in fact, will likely
fail if attempted). Use the path basename as a backup when no
package.json exists.